drivers: audio dai: Move audio dai API header file to audio folder
- to align with documentation, move the dai header to audio foder
- align CMakeLists.txt, Kconfig, prj.conf
