+++ drivers/audio/dai/CMakeLists.txt
@@ -11,6 +11,6 @@ add_subdirectory_ifdef(CONFIG_DAI_NXP_SAI nxp/sai)
 # zephyr-keep-sorted-stop
 
 if(CONFIG_DAI_USERSPACE)
-  zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/drivers/dai.h)
+  zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/audio/dai.h)
   zephyr_library_sources(dai_handlers.c)
 endif()
