+++ drivers/audio/dai/intel/dmic/dmic.c
@@ -18,7 +18,7 @@ LOG_MODULE_REGISTER(LOG_DOMAIN);
 #include <zephyr/pm/device.h>
 #include <zephyr/pm/device_runtime.h>
 
-#include <zephyr/drivers/dai.h>
+#include <zephyr/audio/dai.h>
 #include <zephyr/irq.h>
 
 #include "dmic.h"
