+++ drivers/audio/dai/nxp/sai/sai.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/drivers/dai.h>
+#include <zephyr/audio/dai.h>
 #include <zephyr/device.h>
 #include <zephyr/kernel.h>
 #include <zephyr/pm/device_runtime.h>
